refactor: improve API module structure and remove duplicate code
- Centralize API call logic
- Integrate common fetch configuration and error handling
- Improve date handling for different environments
