+++ src/components/Protest/ProtestMapMarker.tsx
@@ -1,10 +1,10 @@
-import VerificationNumber from '@/lib/API/VerificationNumber';
 import { useEffect, useState } from 'react';
 import { CustomOverlayMap, MapMarker } from 'react-kakao-maps-sdk';
 import { ProtestData } from '@/types';
 import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 import { VerificationBadge } from '@/components/Protest/VerificationBadge';
 import { ProtestCheerBadge } from '@/components/Protest/ProtestCheerBadge';
+import { getVerificationNumber } from '@/apis/verification';
 
 export default function ProtestMapMarker({
     protest,
@@ -16,14 +16,10 @@ export default function ProtestMapMarker({
     router: AppRouterInstance;
 }) {
     const [verifiedNumber, setVerifiedNumber] = useState(0);
-    const date = new Date().toISOString().split('T')[0];
 
     useEffect(() => {
         const verifyNumber = async () => {
-            const result = await VerificationNumber({
-                protestId: protest.id,
-                date: date,
-            });
+            const result = await getVerificationNumber(protest.id);
             setVerifiedNumber(result.verifiedNum);
         };
         verifyNumber();
