+++ src/hooks/SendCheerMutation.ts
@@ -1,0 +1,16 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+export const SendCheerMutation = (protestId: string) => {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: async (protestId: string) =>
+            await fetch(`${process.env.NEXT_PUBLIC_SERVER_DEV_URL}/api/cheer/protest/${protestId}`, {
+                method: 'POST',
+            }),
+        onSuccess: () => {
+            console.log('보냄');
+            queryClient.invalidateQueries({ queryKey: ['cheer', protestId] });
+        },
+    });
+};
